edges, vertexes, and the graph are now managed by the board class
